Minor refactor to support easier modularity
+ start of profile pages
