+++ src/pages/profile/Profile.tsx
@@ -1,0 +1,5 @@
+import { User } from "../../auth/authModule";
+
+export function Profile(context: { user: User }) {
+  return `${context.user.displayName}'s profile`;
+}
