+++ src/pages/profile/ProfileSettings.tsx
@@ -1,0 +1,5 @@
+import { User } from "../../auth/authModule";
+
+export function ProfileSettings(context: { user: User }) {
+  return `${context.user.displayName}'s settings`;
+}
